0.0.0.17.2
+ Структура армии
+ Структура солдат
+ 3 вида солдат
- Изменены интерфейсы и добавлен новый
- Изменена структура экономики и строительства под армию
- Небольшие правки

+++ Assets/Scripts/Player/Economy/Build.cs
@@ -49,8 +49,8 @@ public class Build : MonoBehaviour
     private bool verifyCost(int BuildingId)
     {
         int step = 0;
-        int[] playerRes = GetComponent<Economy>().PlayerEconomy.GetResources();
-        int[] cost = World.GetComponent<WorldList>().BuildingsList[BuildingId].GetCost();
+        float[] playerRes = GetComponent<Economy>().PlayerEconomy.GetResources();
+        float[] cost = World.GetComponent<WorldList>().BuildingsList[BuildingId].GetCost();
 
         for (int i = 0; i < 3; i++)
         {
